refactor: 금액 관련 타입 모두 long 으로 변경

+++ src/main/java/com/hotdeal/reservation/checkout/CheckoutResponse.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 
 public record CheckoutResponse(
         String productName,
-        int price,
+        long price,
         LocalDateTime checkInAt,
         LocalDateTime checkOutAt,
         long pointBalance
